+++ applications/spectrum_analyzer/spectrum_analyzer_worker.h
@@ -4,7 +4,7 @@
 
 typedef void (*SpectrumAnalyzerWorkerCallback)(
     void* chan_table,
-    float max_rssi,
+    double max_rssi,
     uint8_t max_rssi_dec,
     uint8_t max_rssi_channel,
     void* context);
